Only retrieve a single record per unique value for a Solr field
When migrating taxonomy terms, it is not very efficient to go through
all the Islandora records to retrieve some values that are used in
multiple records.
For example, there are only a handfull of genres so it is very
ineffecient to go through all Islandora records to find them.
Luckily Solr has the tools to do this more efficient: facets!
With the unique_field configuration, only 1 Islandora record per
unique value of the field is returned, making the import much
faster and more efficient.